* Moves Pipeline and Switch to sep modules * Updates README and HISTORY
